Move virtual devices code to virtual devices controller
